requirements.txt
streamlit
pandas
numpy
requests
opencv-python
pytesseract
Pillow



